TD-347 "update" API 수정 반영

+++ hbdonf-1116-mk/pages/mypage.js
@@ -4,7 +4,7 @@ import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import { useTranslation } from 'next-i18next'
 import Loading from '../components/Loading';
 import MessageBubble from '../components/MessageBubble';
-import { shareTwitter, shareFacebook, shareLink } from "../lib/util";
+import { shareTwitter, shareFacebook, shareLink, transDate } from "../lib/util";
 import { Icon } from '../components/Icons';
 import Link from 'next/link';
 import { API_URL } from '../lib/config';
@@ -85,12 +85,13 @@ export default function mepage() {
         method: "GET",
         headers: {
             "Content-Type": "application/json",
+            Authorization:fbaseInfo?.accessToken
         },
       }).then((response) => response.json())
       .then((data) => {
-        // console.log(data);
-        if(data.messageId){
-          setMymessage(data);
+        // console.log(data.payload);
+        if(data.payload.messageId){
+          setMymessage(data.payload);
         }
         setLoading(false);
     
@@ -106,10 +107,15 @@ export default function mepage() {
         method: "PUT",
         headers: {
           "Content-Type": "application/json",
+          Authorization:fbaseInfo?.accessToken
         }
-      }).then(()=>{
-        setMymessage();
-        localStorage.clear();
+      }).then((res)=>{
+        if(res.status===200){
+          setMymessage();
+          localStorage.clear();
+        } else{
+          alert(`${res.status} error!`);
+        }
       });
     }
   };
@@ -125,7 +131,7 @@ export default function mepage() {
               <div className='section-my-message'>
                 <p className='user-info'>
                   <b>@{mymessage?.tid}</b>
-                  <span>{mymessage?.date}</span>
+                  <span>{transDate(mymessage?.created)} (KST)</span>
                 </p>
                 <MessageBubble size={40} level={mymessage?.level} text={mymessage?.content} />
               </div>
